+++ Remote/src/Dap.Remote.Aws/Const.fs
@@ -1,0 +1,5 @@
+[<AutoOpen>]
+module Dap.Remote.Aws.Const
+
+[<Literal>]
+let AwsTokenLuid = "AwsToken"