+++ Local/src/Dap.Local.Farango/RELEASE_NOTES.md
@@ -1,2 +1,5 @@
+## 0.6.1
+* Add Document.getAsync helper
+
 ## 0.6.0
 * Libs for develop local applications